Revert "Move coordinators together into a package of their own"
This reverts commit 40dd683b
