+++ reasoner/BUILD
@@ -34,7 +34,6 @@ native_java_libraries(
         ["*.java",
         "resolution/*.java",
         "resolution/framework/*.java",
-        "resolution/resolver/coordinator/*.java",
         "resolution/resolver/*.java",
         "resolution/answer/*.java",
         ], exclude = ["*Test.java", "**/*Test.java"],
